Split views.py
