+++ app/models/__init__.py
@@ -1,5 +1,5 @@
-from .quiz import Quiz
-from .goal import Goal
-from .difficulty import Difficulty
-from .quiz_data import QuizData
-from .player import Player
+from app.models.quiz import Quiz
+from app.models.goal import Goal
+from app.models.difficulty import Difficulty
+from app.models.quiz_data import QuizData
+from app.models.player import Player
