+++ app/views/__init__.py
@@ -1,0 +1,9 @@
+from .answer import *
+from .dif_sel import *
+from .go_goal import *
+from .hints import *
+from .last import *
+from .on_goal import *
+from .opening import *
+from .progress_error import *
+from .reset import *
